Share, Alerts, Verifications put in place.

+++ app/src/main/java/com/example/bjanash_c196/edit_term_view.java
@@ -3,7 +3,6 @@ package com.example.bjanash_c196;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.TextView;
 
@@ -14,12 +13,12 @@ import androidx.appcompat.widget.Toolbar;
 import com.example.bjanash_c196.database.AppDatabase;
 import com.example.bjanash_c196.database.CourseEntity;
 import com.example.bjanash_c196.database.TermEntity;
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.Calendar;
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
 
 public class edit_term_view extends AppCompatActivity {
     TextView TermTitle;
@@ -27,11 +26,11 @@ public class edit_term_view extends AppCompatActivity {
     TextView TermStart;
     Intent intent;
     int termId;
+    int courseId;
     AppDatabase db;
     SimpleDateFormat formatter;
     TermEntity selectedTerm;
-   // Button saveTermButton = findViewById(R.id.termSaveButton);
-   // Button deleteTermButton;
+    List<CourseEntity> allCourses = new ArrayList<>();
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -40,7 +39,11 @@ public class edit_term_view extends AppCompatActivity {
         db = AppDatabase.getInstance(getApplicationContext());
         intent = getIntent();
         setTitle("Edit Term");
+
         termId = intent.getIntExtra("termId", -1);
+        allCourses = db.courseDao().getCourseList(termId);
+        System.out.println(allCourses);
+
         System.out.println(termId);
         selectedTerm = db.termDao().getTerm(termId);
         formatter = new SimpleDateFormat("MM/dd/yyyy HH:mm");
@@ -55,6 +58,7 @@ public class edit_term_view extends AppCompatActivity {
         saveTermButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
+
                 try {
                     String termTitle = TermTitle.getText().toString();
                     String termStartDate = TermStart.getText().toString();
@@ -74,20 +78,23 @@ public class edit_term_view extends AppCompatActivity {
                     e.printStackTrace();
                 }
             }
+
         });
 
         Button deleteTermButton = findViewById(R.id.termDeleteButton);
         deleteTermButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
+                if(allCourses.isEmpty()){
                     TermEntity updatedTerm = db.termDao().getTerm(termId);
                     db.termDao().deleteTerm(updatedTerm);
                     Intent intent = new Intent(getApplicationContext(), list_of_terms.class);
 
-                    startActivity(intent);
-
+                    startActivity(intent);}
+                else {System.out.println("No Courses Attached to This Term -- DELETE");}
             }
-        });
+        }
+                    );
 
 
         Toolbar toolbar = findViewById(R.id.toolbar);
@@ -95,6 +102,8 @@ public class edit_term_view extends AppCompatActivity {
 
 
     }
+
+
 
     private void updateViews() {
         if (selectedTerm != null) {
